+++ src/main/java/Application.java
@@ -1,6 +1,7 @@
 import domain.AutoLottoGenerator;
 import domain.Lotto;
 import domain.LottoGame;
+import domain.LottoTicketOffice;
 import domain.ManualLottoGenerator;
 import domain.Money;
 import java.util.List;
@@ -8,8 +9,6 @@ import view.InputView;
 import view.OutputView;
 
 public class Application {
-
-    private static final String MANUAL_NUMBER_TOO_LARGE_EXCEPTION = "[ERROR] 수동 로또 개수가 너무 많습니다.";
 
     private static Lotto createManualLotto() {
         List<Integer> lottoNumbers = InputView.askLottoNumbers();
@@ -34,11 +33,11 @@ public class Application {
 
     public static void main(String[] args) {
         Money money = new Money(InputView.askMoneyInput());
-        int totalNumber = money.money() / LottoGame.LOTTO_PRICE;
         LottoGame lottoGame = LottoGame.startLottoGame();
+        int totalNumber = LottoTicketOffice.buyLottoTicket(money);
         int manualNumber = InputView.askNumberOfManualLotto();
-        checkValidManualNumber(totalNumber, manualNumber);
-        int autoNumber = totalNumber - manualNumber;
+        int autoNumber = LottoTicketOffice.calculateAutoNumber(totalNumber, manualNumber);
+
         InputView.askManualLottoNumbers();
         clickManualLottoGenerate(lottoGame, manualNumber);
         clickAutoLottoGenerate(lottoGame, autoNumber);
@@ -47,11 +46,5 @@ public class Application {
         lottoGame.enterWinningLottoNumbersAndBonusNumber(InputView.askWinningNumbers(), InputView.askBonusNumber());
         OutputView.printWinningStatistic(lottoGame.produceResults());
         OutputView.printYield(lottoGame.calculateYield());
-    }
-
-    private static void checkValidManualNumber(int totalNumber, int manualNumber) {
-        if (manualNumber > totalNumber) {
-            throw new IllegalArgumentException(MANUAL_NUMBER_TOO_LARGE_EXCEPTION);
-        }
     }
 }
